// discord/topnsfwposts/types/TopListType.ts
interface TopList {
    subreddit: string;
    title: string;
    subreddit_name_prefixed: string;
    ups: number;
    thumbnail: string;
    url_overridden_by_dest: string;
    url: string;
    name: string;
    post_hint: string;
    domain: string;
    id: string;
    permalink: string;
}


export default TopList;